+++ packages/nova-react-test-utils/src/apollo/utils.test.tsx
@@ -7,8 +7,8 @@ import type { EntityCommand, EventWrapper } from "@nova/types";
 import { graphql, useLazyLoadQuery } from "@nova/react";
 
 import { createMockEnvironment, MockPayloadGenerator } from "./test-utils";
-import type { NovaMockEnvironment } from "./nova-mock-environment";
-import { NovaMockEnvironmentProvider } from "./nova-mock-environment";
+import type { NovaMockEnvironment } from "../shared/nova-mock-environment";
+import { NovaMockEnvironmentProvider } from "../shared/nova-mock-environment";
 import { getOperationName, getOperationType } from "./operation-utils";
 
 const schema = buildASTSchema(
