+++ packages/nova-react-test-utils/src/index.ts
@@ -1,20 +1,22 @@
-export { NovaMockEnvironmentProvider } from "./nova-mock-environment";
-export type { NovaMockEnvironment } from "./nova-mock-environment";
-export { MockPayloadGenerator, createMockEnvironment } from "./test-utils";
-export { getOperationName, getOperationType } from "./operation-utils";
-
+export { NovaMockEnvironmentProvider } from "./shared/nova-mock-environment";
+export type { NovaMockEnvironment } from "./shared/nova-mock-environment";
+export { prepareStoryContextForTest } from "./shared/storybook-nova-decorator-shared";
 export type {
   WithNovaEnvironment,
   UnknownOperation,
-} from "./storybook-nova-decorator-shared";
+} from "./shared/storybook-nova-decorator-shared";
+
+
+export { MockPayloadGenerator, createMockEnvironment } from "./apollo/test-utils";
+export { getOperationName, getOperationType } from "./apollo/operation-utils";
 export {
   getNovaApolloDecorator,
   getNovaApolloEnvironmentForStory,
-} from "./storybook-nova-decorator-apollo";
-export { prepareStoryContextForTest } from "./storybook-nova-decorator-shared";
+} from "./apollo/storybook-nova-decorator-apollo";
+
 
 export {
   getNovaRelayDecorator,
   RelayMockPayloadGenerator,
   getNovaRelayEnvironmentForStory,
-} from "./storybook-nova-decorator-relay";
+} from "./relay/storybook-nova-decorator-relay";
